Move content generator into generators folder
