+++ packages/materialized-views/src/generate/content.js
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs';
 import eachSeries from 'async/eachSeries.js';
-import getAllTenants from './utils/get-all-tenants.js';
-import createRepos from './mongodb/create-repos.js';
+import getAllTenants from '../utils/get-all-tenants.js';
+import createRepos from '../mongodb/create-repos.js';
 
 const getRealms = async (repos) => {
   const values = await repos.eventObject.distinct({ key: 'realm' });
