+++ packages/materialized-views/src/index.js
@@ -1,3 +1,3 @@
 export { filterMongoUri } from '@parameter1/events-repositories';
-export { default as generateContent } from './generate-content.js';
+export { default as generateContent } from './generate/content.js';
 export { default as mongodbClient } from './mongodb/client.js';
